move code to where it's used
