+++ apps/web/utils/digest/index.ts
@@ -2,10 +2,7 @@ import { env } from "@/env";
 import { publishToQstashQueue } from "@/utils/upstash";
 import { getCronSecretHeader } from "@/utils/cron";
 import { createScopedLogger } from "@/utils/logger";
-import prisma from "@/utils/prisma";
 import type { DigestBody } from "@/app/api/ai/digest/validation";
-import { DigestStatus } from "@prisma/client";
-import type { DigestEmailSummarySchema } from "@/app/api/resend/digest/validation";
 import type { EmailForAction } from "@/utils/ai/types";
 
 const logger = createScopedLogger("digest");
@@ -47,109 +44,5 @@ export async function enqueueDigestItem({
       emailAccountId,
       error,
     });
-  }
-}
-
-async function findOldestUnsentDigest(emailAccountId: string) {
-  return prisma.digest.findFirst({
-    where: {
-      emailAccountId,
-      status: DigestStatus.PENDING,
-    },
-    orderBy: {
-      createdAt: "asc",
-    },
-  });
-}
-
-export async function upsertDigest({
-  messageId,
-  threadId,
-  emailAccountId,
-  actionId,
-  coldEmailId,
-  content,
-}: {
-  messageId: string;
-  threadId: string;
-  emailAccountId: string;
-  actionId?: string;
-  coldEmailId?: string;
-  content: DigestEmailSummarySchema;
-}) {
-  const logger = createScopedLogger("upsert-digest").with({
-    messageId,
-    threadId,
-    emailAccountId,
-    actionId,
-    coldEmailId,
-  });
-
-  try {
-    // Find or create the digest atomically with digestItems included
-    const digest =
-      (await prisma.digest.findFirst({
-        where: {
-          emailAccountId,
-          status: DigestStatus.PENDING,
-        },
-        orderBy: {
-          createdAt: "asc",
-        },
-        include: {
-          items: {
-            where: {
-              messageId,
-              threadId,
-            },
-            take: 1,
-          },
-        },
-      })) ||
-      (await prisma.digest.create({
-        data: {
-          emailAccountId,
-          status: DigestStatus.PENDING,
-        },
-        include: {
-          items: {
-            where: {
-              messageId,
-              threadId,
-            },
-            take: 1,
-          },
-        },
-      }));
-
-    const digestItem = digest.items.length > 0 ? digest.items[0] : null;
-    const contentString = JSON.stringify(content);
-
-    if (digestItem) {
-      logger.info("Updating existing digest");
-      await prisma.digestItem.update({
-        where: { id: digestItem.id },
-        data: {
-          content: contentString,
-          ...(actionId ? { actionId } : {}),
-          ...(coldEmailId ? { coldEmailId } : {}),
-        },
-      });
-    } else {
-      logger.info("Creating new digest");
-      await prisma.digestItem.create({
-        data: {
-          messageId,
-          threadId,
-          content: contentString,
-          digestId: digest.id,
-          ...(actionId ? { actionId } : {}),
-          ...(coldEmailId ? { coldEmailId } : {}),
-        },
-      });
-    }
-  } catch (error) {
-    logger.error("Failed to upsert digest", { error });
-    throw error;
   }
 }
